pyproject.toml
[tool.poetry]
name = "comecando_fastapi"
version = "0.1.0"
description = ""
authors = ["everton <everton_s.d@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {version = "^0.63.0", extras = ["all"]}
Kivy = "^2.0.0"
kivymd = "^0.104.1"
six = "^1.15.0"
websocket-client = "^0.58.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
